refactor: Huge improvement in folder tree logic
Now the "Extension" check and "File Path" check (from File Contents Export tab) are only run once, after the Folde Tree is completed the check/uncheck action.
Previously it was for EVERY SINGLE check-uncheck WHILE the process was going. So if we checked in a parent folder that had 1000 child folders, then it went through all 1000 children folders to check them too, and do 1000 update for Extensions and another 1000 update for File Path.
Now they only do 1 update for Extension and File Path AFTER the Folder Tree finsihed it's process.